Extracted AutoFixture Customization to public
class, because it enables reuse from e.g. an [InlineAutoData] attribute
(coming up next).
